+++ lib/asn1.c
@@ -75,7 +75,8 @@ shishi_der2asn1 (Shishi * handle,
 
 int
 shishi_a2d_field (Shishi * handle,
-		  Shishi_asn1 node, const char *field, char *der, int *len)
+		  Shishi_asn1 node, const char *field,
+		  char *der, size_t *len)
 {
   char errorDescription[MAX_ERROR_DESCRIPTION_SIZE] = "";
   int rc;
@@ -92,7 +93,7 @@ shishi_a2d_field (Shishi * handle,
 }
 
 int
-shishi_a2d (Shishi * handle, Shishi_asn1 node, char *der, int *len)
+shishi_a2d (Shishi * handle, Shishi_asn1 node, char *der, size_t *len)
 {
   return shishi_a2d_field (handle, node, "", der, len);
 }
